2025.02.17 - 1: removed most Handlers, modernised the data models if the API

+++ app/src/kotlin/ir/mahdiparastesh/instatools/view/Expandable.kt
@@ -76,7 +76,7 @@ class Expandable(
                             nearest(Versioned.BEST),
                             thumb(),
                             media_type.toInt().toByte(),
-                            video_duration?.toLong(),
+                            video_duration?.toInt(),
                             caption?.text
                         )
                     )
